Module Tjr_minifs__Lwt_util

Implement "co-eta" to extract a value from the lwt monad.

We assume that the Lwt main loop is running in a separate Stdlib.Thread. We can then use Lwt_preemptive.run_in_main to run the lwt code and return the result.

val co_eta : Tjr_monad.With_lwt.lwt Tjr_minifs.Minifs_intf.co_eta